+++ CHANGELOG.md
@@ -62,7 +62,7 @@
     - Option type within string interpolation / parameter expansion
     - nested job operator
 - reimplement ``Map`` object. now always preserve insertion order
-- cancel glob or brace expansion in source statement
+- cancel glob expansion in source statement
 
 #### Builtin